requirements.txt
llama-cpp-python>=0.3.0
psutil>=5.9
pytest>=8.0
ruff>=0.4
